ajout de la mécanique de modale pour le batiment music city
